Built-in funtions

Built-in function

Scheduler functions

Agents.as_added — Function.

as_added(model::AbstractModel)

Activates all agents at each step as they have been added to the model.

source

Agents.random_activation — Function.

random_activation(model::AbstractModel)

Activates all agents randomly at each step.

source

Space functions

grid
gridsize
add_agent_to_grid!
move_agent_on_grid!
add_agent_to_grid_single!
move_agent_on_grid_single!
find_empty_nodes_coords
find_empty_nodes
coord_to_vertex
vertex_to_coord
get_node_contents
id_to_agent
node_neighbors

Model functions

Agents.nagents — Function.

nagents(model::AbstractModel)

Returns the number of agents.

source

Agents.kill_agent! — Function.

kill_agent!(agent::AbstractAgent, model::AbstractModel)

Removes an agent from the list of agents and from the space.

source

Agents.step! — Function.

step!(agent_step::Function, model::AbstractModel)

Updates agents one step. Agents will be updated as specified by the model.scheduler.

source

step!(agent_step::Function, model::AbstractModel, nsteps::Integer)

Repeats the step function nsteps times without collecting data.

source

step!(agent_step::Function, model::AbstractModel, nsteps::Integer, agent_properties::Array{Symbol}, steps_to_collect_data::Array{Int64})

Repeats the step function nsteps times, and collects all agent fields in agent_properties at steps steps_to_collect_data.

source

step!(agent_step::Function, model::AbstractModel, nsteps::Integer, agent_properties::Array{Symbol}, aggregators::Array{Function}, steps_to_collect_data::Array{Integer})

Repeats the step function nsteps times, and applies functions in aggregators to values of agent fields in agent_properties at steps steps_to_collect_data.

source

step!(agent_step::Function, model_step::Function, model::AbstractModel)

Updates agents one step without collecting data. This function accepts two functions, one for update agents and one for updating the whole model one after all the agents have been updated.

source

step!(agent_step::Function, model_step::Function, model::AbstractModel, nsteps::Integer)

Repeats the step function nsteps times without collecting data.

source

step!(agent_step::Function, model_step::Function, model::AbstractModel, nsteps::Integer, agent_properties::Array{Symbol}, steps_to_collect_data::Array{Integer})

Repeats the step function nsteps times, and collects all agent fields in agent_properties at steps steps_to_collect_data.

source

step!(agent_step::Function, model_step::Function, model::AbstractModel, nsteps::Integer, agent_properties::Array{Symbol}, aggregators::Array{Function}, steps_to_collect_data::Array{Integer})

Repeats the step function nsteps times, and applies functions in aggregators to values of agent fields in agent_properties at steps steps_to_collect_data.

source