Developer Docs

Internal infrastructure overview

When it comes to development of new code, the overwhelming majority of Agents.jl is composed of three parts:

The model time-stepping dynamics and agent storage and retrieval logic

The space agent storage, movement, and neighborhood searching logic

The rest of the API which is largely agnostic to the above two

Arguably the most important aspect of the Agents.jl design is that the above three pillars of the infrastructure are orthogonal. That is, if someone wanted to add a new space, they would not have to care about neither the model stepping dynamics, nor the majority of the remaining Agents.jl API, such as sampling, data collection, etc. etc.

Cloning the repository and formatting

Since we include documentation with many animated gifs and videos in the repository, a standard clone can be larger than expected. If you wish to do any development work, it is better to use

git clone https://github.com/JuliaDynamics/Agents.jl.git --single-branch

Agents.jl uses the Runic.jl formatter. Make sure your contributions apply it! To apply the formatter use

shell> julia --project=@runic --startup-file=no -e 'using Pkg; Pkg.add("Runic")'
shell> julia --project=@runic --startup-file=no -e 'using Runic; exit(Runic.main(ARGS))' -- --inplace path/to/developed/Agents

Creating a new model type

Note that new model types target applications where a fundamentally different way to define the "time evolution" or "dynamic rule" is required. If any of the existing AgentBasedModel subtypes satisfy the type of time stepping, then you don't have to create a new type of model.

Creating a new model type within Agents.jl is simple. Although it requires extending a bit more than a dozen functions, the majority of them are 1-liner "accessor" functions (that return e.g., the rng, or the space instance).

The most important mandatory method is to extend step! for your new type. You can see the existing implementations of step! for e.g., StandardABM or EventQueueABM to get an idea.

All other mandatory method extensions for e.g., accessor functions are in the file src/core/model_abstract.jl. As you will notice, the overwhelming majority of required methods have a default implementation that e.g., attempts to return a field named :rng. The rest of the methods by default return a "not implemented" error message (and those you also need to extend mandatorily).

Creating a new space type

Creating a new space type within Agents.jl is quite simple and requires the extension of only 5 methods to support the entire Agents.jl API. Here are the steps to follow to create a new space:

Think about what the agent position type should be.

Think about how the space type will keep track of the agent positions, so that it is possible to implement the function nearby_ids.

Create the struct that represents your new space, while making it a subtype of Agents.AbstractSpace. Let now const ABMS = ABM{<:YourSpaceType}.

Extend random_position(model::ABMS).

Extend add_agent_to_space!(agent, model::ABMS), remove_agent_from_space!(agent, model::ABMS). This already provides access to add_agent!, kill_agent! and move_agent!.

Extend nearby_ids(pos, model::ABMS, r; kw...).

Create a new "minimal" agent type to be used with @agent (see the source code of e.g., GraphAgent for an example).

And that's it! Every function of the main API will now work. In some situations you might want to explicitly extend other functions such as move_agent! or remove_all_from_space! for performance reasons, but they will work out of the box with a generic implementation.

Visualization of a custom space

Visualization of a new space type within Agents.jl works in a similar fashion to creating a new space type. One your space works with the general Agents.jl API, you only need to extend a few functions for it to work automatically with the existing plotting and animation infrastructure.

Mandatory methods

You must extend the following function

Agents.space_axis_limits — Function

space_axis_limits(model::ABM{S}) where {S<:AbstractSpace}

Return appropriate axis limits for given model. They have to be a 2- or 3- tuple as expected by Makie when setting axis limits. The length of the output of this function also decides whether a 2D or 3D axis is used.

Optional alternative agent plotting

If your space does not visualize agents in the default way of one agent = one scattered marker, then you want to extend the function agentsplot!. For example, GraphSpace aggregates multiple agents into a graph plot and is extending agentsplot!.

Agents.agentsplot! — Function

agentsplot!(ax, model, agent_color, agent_size, agent_marker, offset, agentsplotkwargs)

Plot agents into ax, given model and agent_color, agent_size, agent_marker, offset, agentsplotkwargs which are obtained from the keywords given to abmplot. The model is the model observable contained in ABMObservable.

Extending

By default this function does a scatter-plot and ignores model. Otherwise, it should be extended to dispatch on model space type. TO do this, use the signature:

agentsplot!(ax, model::T, args...) where {T <: Observable{A} where {A <: ABM{<:YouSpaceType}}}

See the source code of the default implementation of agentsplot! for lifting e.g., the agent_color argument into the model.

Optional pre-plotting

Some spaces, like the OSMSpace, need to plot some elements before the agents can be plotted and animated. If that's the case, extend the following:

Agents.spaceplot! — Function

spaceplot!(ax, space::AbstractSpace; spaceplotkwargs...)

Add a space-dependent plot to ax. spaceplotkwargs is propagated from abmplot. By default this function does nothing.

Heatmap handling

Heatmaps are extracted and plotted automatically, but your space may require some special handling for that. For example ContinuousSpace needs to map the finite heatmap matrix over the continuous space. If you require such handling, extend:

Agents.abmheatmap! — Function

abmheatmap!(ax, abmobs::ABMObservable, space, heatobs, add_colorbar, heatkwargs)

Plot as a heatmap the provided heatobs (an observable matrix created from heatarray). Return the heatmap (or otherwise) plot object so that a Colorbar can be assigned to it.

Designing a new Pathfinder Cost Metric

To define a new cost metric, simply make a struct that subtypes CostMetric and provide a delta_cost function for it. These methods work solely for A* at present, but will be available for other pathfinder algorithms in the future.

Agents.Pathfinding.CostMetric — Type

Pathfinding.CostMetric{D}

An abstract type representing a metric that measures the approximate cost of travelling between two points in a D dimensional grid.

Agents.Pathfinding.delta_cost — Function

Pathfinding.delta_cost(pathfinder::GridPathfinder{D}, metric::M, from, to) where {M<:CostMetric}

Calculate an approximation for the cost of travelling from from to to (both of type NTuple{N,Int}. Expects a return value of Float64.

Implementing custom serialization

For model properties

Custom serialization may be required if your properties contain non-serializable data, such as functions. Alternatively, if it is possible to recalculate some properties during deserialization it may be space-efficient to not save them. To implement custom serialization, define methods for the to_serializable and from_serializable functions:

Agents.AgentsIO.to_serializable — Function

AgentsIO.to_serializable(t)

Return the serializable form of the passed value. This defaults to the value itself, unless a more specific method is defined. Define a method for this function and for AgentsIO.from_serializable if you need custom serialization for model properties. This also enables passing keyword arguments to AgentsIO.load_checkpoint and having access to them during deserialization of the properties. Some possible scenarios where this may be required are:

Your properties contain functions (or any type not supported by JLD2.jl). These may not be (de)serialized correctly. This could result in checkpoint files that cannot be loaded back in, or contain reconstructed types that do not retain their data/functionality.

Your properties contain data that can be recalculated during deserialization. Omitting such properties can reduce the size of the checkpoint file, at the expense of some extra computation at deserialization.

If your model properties do not fall in the above scenarios, you do not need to use this function.

This function, and AgentsIO.from_serializable is not called recursively on every type/value during serialization. The final serialization functionality is enabled by JLD2.jl. To define custom serialization for every occurrence of a specific type (such as agent structs), refer to the Custom Serialization section of JLD2.jl documentation.

Agents.AgentsIO.from_serializable — Function

AgentsIO.from_serializable(t; kwargs...)

Given a value in its serializable form, return the original version. This defaults to the value itself, unless a more specific method is defined. Define a method for this function and for AgentsIO.to_serializable if you need custom serialization for model properties. This also enables passing keyword arguments to AgentsIO.load_checkpoint and having access to them through kwargs.

Refer to AgentsIO.to_serializable for more info.

For agent structs

Similarly to model properties, you may need to implement custom serialization for agent structs. from_serializable and to_serializable are not called during (de)serialization of agent structs. Instead, JLD2's custom serialization functionality should be used. All instances of the agent struct will be converted to and from the specified type during serialization. For OpenStreetMap agents, the position, destination and route are saved separately. These values will be loaded back in during deserialization of the model and override any values in the agent structs. To save space, the agents in the serialized model will have empty route fields.

OpenStreetMapSpace internals

Details about the internal details of the OSMSpace are discussed in the docstring of OSM.OpenStreetMapPath.

Benchmarking

As Agents.jl is developed we want to monitor code efficiency through benchmarks. A benchmark is a function or other bit of code whose execution is timed so that developers and users can keep track of how long different API functions take when used in various ways. Individual benchmarks can be organized into suites of benchmark tests. See the benchmark directory to view Agents.jl's benchmark suites. Follow these examples to add your own benchmarks for your Agents.jl contributions. See the BenchmarkTools quickstart guide, toy example benchmark suite, and the BenchmarkTools.jl manual for more information on how to write your own benchmarks.

Creating a new AgentBasedModel implementation

The interface defined by AgentBasedModel, that needs to be satisfied by new implementations, is very small. It is contained in the file src/core/model_abstract.jl.