API

Model and space information

LightGraphs.nv — Method

nv(model::ABM)

Return the number of nodes (vertices) in the model space.

LightGraphs.ne — Method

ne(model::ABM)

Return the number of edges in the model space.

Agents.has_empty_nodes — Function

has_empty_nodes(model)

Return true if there are empty nodes in the model.

Agents.find_empty_nodes — Function

find_empty_nodes(model::ABM)

Returns the IDs of empty nodes on the model space.

Content from a node

Agents.node_neighbors — Function

node_neighbors(agent::AbstractAgent, model::ABM)
node_neighbors(node::Int, model::ABM)

Return neighboring node coordinates/numbers of the node on which the agent resides.

If the model's space is GraphSpace, then the function will return node numbers. If space is GridSpace then the neighbors are returned as coordinates (tuples).

node_neighbors(node_number::Integer, model::ABM, radius::Integer)

Returns a list of neighboring nodes to the node node_number within the radius. radius defines higher degree neighbors. For example, neighbors with a radius=2 include first and second degree neighbors, that is, neighbors and neighbors of neighbors.

Notice that node_neighbors excludes the node that is given.

Agents.pick_empty — Function

pick_empty(model)

Return the ID of a random empty node or 0 if there are no empty nodes.

Agents.get_node_contents — Function

get_node_contents(n::Integer, model)

Return the ids of agents in the node n of the model.

get_node_contents(agent::AbstractAgent, model)

Return all agents' ids in the same node as the agent (including the agent's own id).

get_node_contents(coords::Tuple, model)

Return the ids of agents in the node at coords.

Base.isempty — Method

isempty(node::Int, model::ABM)

Return true if there are no agents in node.

Agent information and retrieval

Agents.nagents — Function

nagents(model::ABM)

Return the number of agents in the model.

Agents.id2agent — Function

id2agent(id::Integer, model)

Return an agent given its ID.

Agents.random_agent — Function

random_agent(model)

Return a random agent from the model.

Model-Agent interaction

Agents.add_agent! — Function

add_agent!(agent::AbstractAgent [, pos], model::ABM) → agent

Adds the agent to the pos in the space and to the list of agents. If pos is not given, the agent is added to a random position. The agent's position is always updated to match pos.

add_agent!(node, model::ABM, properties...)

Add a new agent in the given node, by constructing the agent type of the model and propagating all extra properties to the constructor.

add_agent!(model::ABM, properties...)

Similar with add_agent!(node, model, properties...), but adds the created agent to a random node. This function also works for models without a spatial structure.

Agents.add_agent_single! — Function

add_agent_single!(agent::AbstractAgent, model::ABM) → agent

Add agent to a random node in the space while respecting a maximum one agent per node. This function does not do anything if there are no empty nodes.

Return the new agent.

add_agent_single!(model::ABM, properties...)

Same as add_agent!(model, properties...) but ensures that it adds an agent into a node with no other agents (does nothing if no such node exists).

Agents.move_agent! — Function

move_agent!(agent::AbstractAgent [, pos], model::ABM) → agent

Add agentID to the new position pos (either tuple or integer) in the model and remove it from the old position (also update the agent to have the new position).

If pos is a tuple, it represents the coordinates of the grid node. If pos is an integer, it represents the node number in the graph. If pos is not given, the agent is moved to a random position on the grid.

Agents.move_agent_single! — Function

move_agent_single!(agent::AbstractAgent, model::ABM)

Move agent to a random nodes on the grid while respecting a maximum of one agent per node. If there are no empty nodes, the agent wont move.

Return the agent's new position.

Agents.kill_agent! — Function

kill_agent!(agent::AbstractAgent, model::ABM)

Remove an agent from the list of agents and from the space.

Iteration

Agents.NodeIterator — Type

NodeIterator(model) → iterator

Create an iterator that returns node coordinates, if the space is a grid, or otherwise node number, and the agent IDs in each node.

Agents.nodes — Function

nodes(model; by = :id) -> ns

Return a vector of the node ids of the model that you can iterate over. The ns are sorted depending on by:

:id - just sorted by their number

:random - randomly sorted

:population - nodes are sorted depending on how many agents they accommodate. The more populated nodes are first.

Schedulers

Agents.as_added — Function

as_added(model::ABM)

Activate agents at each step in the same order as they have been added to the model.

Agents.random_activation — Function

random_activation(model::ABM)

Activate agents once per step in a random order.

Agents.partial_activation — Function

partial_activation(model::ABM)

At each step, activate only activation_prob number of randomly chosen of individuals with a activation_prob probability. activation_prob must be a field in the model and between 0 and 1.

Utilities

Agents.coord2vertex — Function

coord2vertex(coord::NTuple{Int}, model_or_space) → n
coord2vertex(coord::AbstractAgent, model_or_space) → n

Return the node number n of the given coordinates or the agent's position.

Agents.vertex2coord — Function

vertex2coord(vertex::Integer, model_or_space) → coords

Returns the coordinates of a node given its number on the graph.

Parallelization

Agents.parallel_replicates — Function

parallel_replicates(agent_step!, model::ABM, n::Integer, agent_properties::Array{Symbol}, when::AbstractArray{Integer}, replicates::Integer)

Runs replicates number of simulations in parallel and returns a DataFrame.