API

The core API is defined by AgentBasedModel, Space, AbstractAgent and step!, which are described in the Tutorial page. The functionality described here builds on top of the core API.

Model and space information

LightGraphs.nv — Method

nv(model::ABM)

Return the number of nodes (vertices) in the model space.

LightGraphs.ne — Method

ne(model::ABM)

Return the number of edges in the model space.

Agents.has_empty_nodes — Function

has_empty_nodes(model)

Return true if there are empty nodes in the model.

Agents.find_empty_nodes — Function

find_empty_nodes(model::ABM)

Returns the IDs of empty nodes on the model space.

Content from a node

Agents.node_neighbors — Function

node_neighbors(agent::AbstractAgent, model::ABM)
node_neighbors(node::Int, model::ABM)

Return neighboring node coordinates/numbers of the node on which the agent resides.

If the model's space is GraphSpace, then the function will return node numbers. If space is GridSpace then the neighbors are returned as coordinates (tuples).

node_neighbors(node_number::Integer, model::ABM, radius::Integer)

Returns a list of neighboring nodes to the node node_number within the radius. radius defines higher degree neighbors. For example, neighbors with a radius=2 include first and second degree neighbors, that is, neighbors and neighbors of neighbors.

Notice that node_neighbors excludes the node that is given.

Agents.pick_empty — Function

pick_empty(model)

Return the ID of a random empty node or 0 if there are no empty nodes.

Agents.get_node_contents — Function

get_node_contents(n::Integer, model)

Return the ids of agents in the node n of the model.

get_node_contents(agent::AbstractAgent, model)

Return all agents' ids in the same node as the agent (including the agent's own id).

get_node_contents(coords::Tuple, model)

Return the ids of agents in the node at coords.

Agents.get_node_agents — Function

get_node_agents(x, model)

Same as get_node_contents(x, model) but directly returns the list of agents instead of just the list of IDs.

Base.isempty — Method

isempty(node::Int, model::ABM)

Return true if there are no agents in node.

Agent information and retrieval

Agents.nagents — Function

nagents(model::ABM)

Return the number of agents in the model.

Agents.id2agent — Function

id2agent(id::Integer, model)

Return an agent given its ID.

Agents.random_agent — Function

random_agent(model)

Return a random agent from the model.

Model-Agent interaction

Agents.add_agent! — Function

add_agent!(agent::AbstractAgent [, pos], model::ABM) → agent

Add the agent to the pos in the space and to the list of agents. If pos is not given, the agent is added to a random position. The agent's position is always updated to match pos.

add_agent!(node, model::ABM, properties...)

Add a new agent in the given node, by constructing the agent type of the model and propagating all extra properties to the constructor.

Notice that this function takes care of setting the agent's id and position and thus properties... is propagated to other fields the agent has.

Example

using Agents
mutable struct Agent <: AbstractAgent
    id::Int
    w::Float64
end
m = ABM(Agent) # model without spatial structure
add_agent!(m, 1, rand()) # incorrect: id is set internally
add_agent!(m, rand()) # correct: weight becomes rand()

add_agent!(model::ABM, properties...)

Similar with add_agent!(node, model, properties...), but adds the created agent to a random node. This function also works for models without a spatial structure.

Agents.add_agent_pos! — Function

add_agent_pos!(agent::AbstractAgent, model::ABM)

Add the agent to the model at the agent's own position if the agent has a position, or to a random position otherwise.

Agents.add_agent_single! — Function

add_agent_single!(agent::AbstractAgent, model::ABM) → agent

Add agent to a random node in the space while respecting a maximum one agent per node. This function does not do anything if there are no empty nodes.

Return the new agent.

add_agent_single!(model::ABM, properties...)

Same as add_agent!(model, properties...) but ensures that it adds an agent into a node with no other agents (does nothing if no such node exists).

Agents.move_agent! — Function

move_agent!(agent::AbstractAgent [, pos], model::ABM) → agent

Add agentID to the new position pos (either tuple or integer) in the model and remove it from the old position (also update the agent to have the new position).

If pos is a tuple, it represents the coordinates of the grid node. If pos is an integer, it represents the node number in the graph. If pos is not given, the agent is moved to a random position on the grid.

move_agent!(agent::A, model::ABM{A, ContinuousSpace})

In the case of continuous space, move_agent! propagates the agent forwards one step according to its velocity, after updating the agent's velocity (see Space).

Agents.move_agent_single! — Function

move_agent_single!(agent::AbstractAgent, model::ABM)

Move agent to a random nodes on the grid while respecting a maximum of one agent per node. If there are no empty nodes, the agent wont move.

Return the agent's new position.

Agents.kill_agent! — Function

kill_agent!(agent::AbstractAgent, model::ABM)

Remove an agent from model, and from the space if the model has a space.

Agents.genocide! — Function

genocide!(model::ABM)

Kill all the agents of the model.

genocide!(model::ABM, n::Int)

Kill the agents of the model whose IDs are larger than n.

genocide!(model::ABM, f::Function)

Kill all agents where the function f(agent) returns true.

Simulations

The central simulation function is step!, which is mentioned in our Tutorial. But there are other functions that are related to simulations listed here.

Agents.paramscan — Function

paramscan(parameters, initialize; kwargs...)

Run the model with all the parameter value combinations given in parameters, while initializing the model with initialize. This function uses DrWatson's dict_list internally. This means that every entry of parameters that is a Vector, contains many parameters and thus is scanned. All other entries of parameters that are not Vectors are not expanded in the scan.

initialize is a function that creates an ABM. It should accept keyword arguments.

Keywords

All the following keywords are propagated into step!: agent_step!, properties, n, when = 1:n, model_step! = dummystep, step0::Bool = true, parallel::Bool = false, replicates::Int = 0.

The following keywords modify the paramscan function:

include_constants::Bool=false determines whether constant parameters should be included in the output DataFrame.

progress::Bool = true whether to show the progress of simulations.

Agents.sample! — Function

sample!(model::ABM, n [, weight]; kwargs...)

Replace the agents of the model with a random sample of the current agents with size n.

Optionally, choose an agent property weight (Symbol) to weight the sampling. This means that the higher the weight of the agent, the higher the probability that this agent will be chosen in the new sampling.

Keywords

replace = true : whether sampling is performed with replacement, i.e. all agents can be chosen more than once.

rng = GLOBAL_RNG : a random number generator to perform the sampling with.

See the Wright-Fisher example in the documentation for an application of sample!.

Iteration

Agents.NodeIterator — Type

NodeIterator(model) → iterator

Create an iterator that returns node coordinates, if the space is a grid, or otherwise node number, and the agent IDs in each node.

Agents.nodes — Function

nodes(model; by = :id) -> ns

Return a vector of the node ids of the model that you can iterate over. The ns are sorted depending on by:

:id - just sorted by their number

:random - randomly sorted

:population - nodes are sorted depending on how many agents they accommodate. The more populated nodes are first.

Schedulers

The schedulers of Agents.jl have a very simple interface. All schedulers are functions, that take as an input the ABM and return an iterator over agent IDs. Notice that this iterator can be a "true" iterator or can be just a standard vector of IDs. You can define your own scheduler according to this API and use it when making an AgentBasedModel.

Agents.fastest — Function

fastest

Activate all agents once per step in the order dictated by the agent's container, which is arbitrary (the keys sequence of a dictionary). This is the fastest way to activate all agents once per step.

Agents.by_id — Function

by_id

Activate agents at each step according to their id.

Agents.random_activation — Function

random_activation

Activate agents once per step in a random order. Different random ordering is used at each different step.

Agents.partial_activation — Function

partial_activation(p)

At each step, activate only p percentage of randomly chosen agents.

Agents.property_activation — Function

property_activation(property)

At each step, activate the agents in an order dictated by their property, with agents with greater property acting first. property is a Symbol, which just dictates which field the agents to compare.

Utilities

Agents.coord2vertex — Function

coord2vertex(coord::NTuple{Int}, model_or_space) → n
coord2vertex(coord::AbstractAgent, model_or_space) → n

Return the node number n of the given coordinates or the agent's position.

Agents.vertex2coord — Function

vertex2coord(vertex::Integer, model_or_space) → coords

Returns the coordinates of a node given its number on the graph.

Parallelization

Agents.parallel_replicates — Function

parallel_replicates(agent_step!, model::ABM, n::Integer, agent_properties::Array{Symbol}, when::AbstractArray{Integer}, replicates::Integer)

Runs replicates number of simulations in parallel and returns a DataFrame.

Plotting

Plotting functionality comes from AgentsPlots, which uses Plots.jl. You need to install both AgentsPlots, as well as a plotting backend (we use GR) to use the following functions.

AgentsPlots.plotabm — Function

plotabm(model::ABM [, c , s]; kwargs...)

Plot the model as a graph by providing two optional functions c, s. Both of these functions gain as an input the list of agents a at a node of the model's graph and output a number. c returns a color (anything acceptable by Plots.jl) of the node while s returhs the (relative) size of the node.

c defaults to a purple color for all nodes, while s defaults to length. Internally the graphplot recipe is used, and thus all keyword arguments are propagated into this recipe.

If you use plotabm in an animation loop, notice that not all method types of graphplot are deterministic and thus you should probably set Random.seed! before each call to plotabm (and call Random.seed!(rand(1:10000)) afterwards to ensure random behavior for your abm).

AgentsPlots.plot2D — Function

plot2D(node_coords::AbstractArray, colors::AbstractArray; kwargs...)

Creates a scatter plot for nodes and their colors

node_coords: node positions as coordinates

colors: color of each node

Keywords

nodesize=1.0

markeralpha = nothing

plot2D(data::AbstractDataFrame, status_column::Symbol; kwargs)

Plots the distribution of agents on a 2D grid. Agent positions should be saved as tuples.

data: A dataframe output of your simulation.

status_column: the name of a column that determines category of each agent, so that it is colored differently.

Keywords

cc::Dict=Dict() Optionally provide a color name for each unique value in the status_column

nodesize=1.0 size of each node

t=0 the time step of the simulation to plot.

AgentsPlots.plot_CA1D — Function

plot_CA1D(data; Keywords)

Visualizes data of a 1D cellular automaton.

data: output of CA1D.ca_run.

Keywords

nodesize=2.0: Size of each cell.

AgentsPlots.plot_CA2D — Function

plot_CA2D(data; kwargs...)

Visualizes data of a 2D cellular automaton.

data: output of CA2D.ca_run.

Keywords

t=nothing : The time step to be plotted. If nothing, all the rows in data are used.

nodesize=2.0

AgentsPlots.plot_CA2Dgif — Function

plot_CA2Dgif(data; kwargs...)

Create a 2D scatter plot from all data and adds a frame to the anim animation object. If anim is not provided, it creates a new one. Returns an animation object. It can be saved as an animated gif using AgentsPlots.gif(anim, "filename.gif").

data: output of of time-step of CA2D.ca_run.

Keywords

anim::Animation=Animation() : animation object. If provided, a new frame is added to it.

nodesize=2.0