Built-in funtions

Built-in function

Scheduler functions

Agents.as_added — Function.

as_added(model::AbstractModel)

Activates agents at each step in the same order as they have been added to the model.

source

Agents.random_activation — Function.

random_activation(model::AbstractModel)

Activates agents once per step in a random order.

source

Agents.partial_activation — Function.

partial_activation(model::AbstractModel)

At each step, activates only activation_prob number of randomly chosen of individuals with a activation_prob probability. activation_prob should be a field in the model and between 0 and 1.

source

Space functions

Agents.grid — Function.

grid(x::Integer, y::Integer, z::Integer, periodic=false, Moore=false)

Return a grid based on its dimensions. x, y, and z are the dimensions of the grid. If all dimensions are 1, it will return a 0D space, where all agents are in the same position. If x is more than 1, but y and z are 1, it will return a 1D grid. If x and y are more than 1, and z=1, it will return a 2D regular grid.

periodic=true will create toroidal grids.

Moore=true will return a regular grid in which each node is connected to its diagonal neighbors. If false, each node will only connect to its orthogonal neighbors.

source

grid(dims::Tuple{Integer, Integer, Integer}, periodic=false, Moore=false)

Return a grid based on its dimensions. x, y, and z are the dimensions of the grid. If all dimensions are 1, it will return a 0D space, where all agents are in the same position. If x is more than 1, but y and z are 1, it will return a 1D grid. If x and y are more than 1, and z=1, it will return a 2D regular grid.

periodic=true will create toroidal grids.

Moore=true will return a regular grid in which each node is connected to its diagonal neighbors. If false, each node will only connect to its orthogonal neighbors.

source

grid(dims::Tuple{Integer, Integer}, periodic=false, Moore=false)

Return a grid based on its dimensions. x, y are the dimensions of the grid. If all dimensions are 1, it will return a 0D space, where all agents are in the same position. If x is more than 1, but y is 1, it will return a 1D grid.

periodic=true will create toroidal grids.

Moore=true will return a regular grid in which each node is connected to its diagonal neighbors. If false, each node will only connect to its orthogonal neighbors.

source

Agents.gridsize — Function.

gridsize(dims::Tuple{Integer, Integer, Integer})

Returns the size of a grid with dimenstions dims.

source

gridsize(dims::Tuple{Integer, Integer})

Returns the size of a grid with dimenstions dims.

source

gridsize(model::AbstractModel)

Returns the size of the grid in the model

source

Agents.Node_iter — Type.

Node_iter(model::AbstractModel)

An iterator that returns node coordinates, if the graph is a grid, or otherwise node numbers, and the agents in each node.

source

Agents.add_agent! — Function.

add_agent!(agent::AbstractAgent, pos::Tuple{Integer, Integer, Integer}, model::AbstractModel)

Adds the agent to the pos in the space and to the list of agents. pos is tuple of x, y, and z (only if its a 3D space) coordinates of the grid node. If pos is not given, the agent is added to a random position.

This function is for positioning agents on the grid for the first time.

source

add_agent!(agent::AbstractAgent, pos::Integer, model::AbstractModel)

Adds the agent to the pos in the space and to the list of agents. pos is the node number of the space. If pos is not given, the agent is added to a random position.

This function is for positioning agents on the grid for the first time.

source

add_agent!(agent::AbstractAgent, model::AbstractModel)

Adds agent to a random node in the space and to the agent to the list of agents.

Returns the agent's new position.

source

Agents.move_agent! — Function.

move_agent!(agent::AbstractAgent, pos::Tuple, model::AbstractModel)

Adds agentID to a new position in the grid and removes it from the old position. Also updates the agent to represent the new position. pos is tuple of x, y, z (only if its a 3D space) coordinates of the grid node. If pos is not given, the agent is moved to a random position on the grid.

source

move_agent!(agent::AbstractAgent, pos::Integer, model::AbstractModel)

Adds agentID to a new position in the grid and removes it from the old position. Also updates the agent to represent the new position. pos is an integer showing the number of the node on the grid node. If pos is not given, the agent is moved to a random position on the grid.

source

Agents.add_agent_single! — Function.

add_agent_single!(agent::AbstractAgent, model::AbstractModel)

Adds agent to a random node in the space while respecting a maximum one agent per node. It does not do anything if there are no empty nodes.

Returns the agent's new position.

source

Agents.move_agent_single! — Function.

move_agent_single!(agent::AbstractAgent, model::AbstractModel)

Moves agent to a random nodes on the grid while respecting a maximum of one agent per node. If there are no empty nodes, the agent wont move.

Return the agent's new position.

source

Agents.empty_nodes — Function.

empty_nodes(model::AbstractArray)

Returns true if there are empty nodes, otherwise returns false.

source

Agents.pick_empty — Function.

pick_empty(model)

Returns the ID of a random empty cell. Returns 0 if there are no empty cells

source

Agents.find_empty_nodes_coords — Function.

find_empty_nodes_coords(model::AbstractModel)

Returns the coordinates of empty nodes on the model grid.

source

Agents.find_empty_nodes — Function.

find_empty_nodes(model::AbstractModel)

Returns the IDs of empty nodes on the model space.

source

Agents.coord_to_vertex — Function.

coord_to_vertex(coord::Tuple{Integer, Integer, Integer}, model::AbstractModel)

Returns the node number from x, y, z coordinates.

source

coord_to_vertex(coord::Tuple{Integer, Integer, Integer}, dims::Tuple{Integer, Integer, Integer})

Returns node number from x, y, z coordinates.

source

Agents.vertex_to_coord — Function.

vertex_to_coord(vertex::Integer, model::AbstractModel)

Returns the coordinates of a node given its number on the graph.

source

vertex_to_coord(vertex::Integer, dims::Tuple{Integer, Integer, Integer})

Returns the coordinates of a node given its number on a 3D grid.

source

vertex_to_coord(vertex::Integer, dims::Tuple{Integer,Integer})

Returns the coordinates of a node given its number on a 2D grid.

source

Agents.get_node_contents — Function.

get_node_contents(agent::AbstractAgent, model::AbstractModel)

Returns all agents' ids in the same node as the agent.

source

get_node_contents(coords::Tuple, model::AbstractModel)

Returns the id of agents in the node at coords

source

get_node_contents(node_number::Integer, model::AbstractModel)

Returns the id of agents in the node at node_number

source

Agents.id_to_agent — Function.

id_to_agent(id::Integer, model::AbstractModel)

Returns an agent given its ID.

source

Agents.node_neighbors — Function.

node_neighbors(agent::AbstractAgent, model::AbstractModel)

Returns neighboring node coords/numbers of the node on which the agent resides. If agent pos is recorded an integer, the function will return node numbers of the neighbors. If the agent pos is a tuple, the function will return the coordinates of neighbors on a grid.

source

node_neighbors(node_number::Integer, model::AbstractModel)

Returns neighboring node IDs of the node with node_number.

source

node_neighbors(node_coord::Tuple, model::AbstractModel)

Returns neighboring node coords of the node with node_coord.

source

node_neighbors(node_number::Integer, model::AbstractModel, radius::Integer)

Returns a list of neighboring cells to the node node_number within the radius.

source

Model functions

Agents.nagents — Function.

nagents(model::AbstractModel)

Returns the number of agents.

source

Agents.kill_agent! — Function.

kill_agent!(agent::AbstractAgent, model::AbstractModel)

Removes an agent from the list of agents and from the space.

source

Agents.step! — Function.

step!(agent_step::Function, model::AbstractModel)

Updates agents one step. Agents will be updated as specified by the model.scheduler.

source

step!(agent_step::Function, model::AbstractModel, nsteps::Integer)

Repeats the step function nsteps times without collecting data.

source

step!(agent_step::Function, model::AbstractModel, nsteps::Integer, agent_properties::Array{Symbol}, steps_to_collect_data::Array{Int64})

Repeats the step function nsteps times, and collects all agent fields in agent_properties at steps steps_to_collect_data.

source

step!(agent_step::Function, model::AbstractModel, nsteps::Integer, agent_properties::Array{Symbol}, aggregators::Array{Function}, steps_to_collect_data::Array{Integer})

Repeats the step function nsteps times, and applies functions in aggregators to values of agent fields in agent_properties at steps steps_to_collect_data.

source

step!(agent_step::Function, model_step::Function, model::AbstractModel)

Updates agents one step without collecting data. This function accepts two functions, one for update agents and one for updating the whole model one after all the agents have been updated.

source

step!(agent_step::Function, model_step::Function, model::AbstractModel, nsteps::Integer)

Repeats the step function nsteps times without collecting data.

source

step!(agent_step::Function, model_step::Function, model::AbstractModel, nsteps::Integer, agent_properties::Array{Symbol}, steps_to_collect_data::Array{Integer})

Repeats the step function nsteps times, and collects all agent fields in agent_properties at steps steps_to_collect_data.

source

step!(agent_step::Function, model_step::Function, model::AbstractModel, nsteps::Integer, agent_properties::Array{Symbol}, aggregators::Array{Function}, steps_to_collect_data::Array{Integer})

Repeats the step function nsteps times, and applies functions in aggregators to values of agent fields in agent_properties at steps steps_to_collect_data.

source

Data collectors

combine_columns!
write_to_file
data_collector

Batch runner

batchrunner

Visualization functions

agents_plots_complete
visualize_data
visualize_2D_agent_distribution
visualize_1DCA
visualize_2DCA