Tutorial

YouTube video

This Tutorial is also available as a YouTube video: https://youtu.be/fgwAfAa4kt0

In Agents.jl a central abstract structure called AgentBasedModel contains all information necessary to run a simulation: the evolution rule (also called dynamic rule), the agents of the simulation, and other global properties relevant to the simulation. AgentBasedModels map unique IDs (integers) to agent instances. During the simulation, the model evolves in discrete steps. During one step, the user decides which agents will act, how they will act, how many times, and whether any model-level properties will be adjusted. Once the time evolution is defined, collecting data during time evolution is straightforward by simply stating which data should be collected.

In the spirit of simple design, all of this is done by defining simple Julia data types, like basic functions, structs and dictionaries.

To set up an ABM simulation in Agents.jl, a user only needs to follow these steps:

Choose what kind of space the agents will live in, for example a graph, a grid, etc. Several spaces are provided by Agents.jl and can be initialized immediately.

Define the agent type (or types, for mixed models) that will populate the ABM. Agent types are Julia mutable structs that are created with @agent. The types must contain some mandatory fields, which is ensured by using @agent. The remaining fields of the agent type are up to the user's choice.

Define the evolution rule, i.e., how the model evolves in time. The evolution rule needs to be provided as at least one, or at most two functions: an agent stepping function, that acts on each agent one by one, and/or a model stepping function, that steps the entire model as a whole. These functions are standard Julia functions that take advantage of the Agents.jl API.

Initialize an agent based model instance that contains created agent type, the chosen space, the evolution rule, other optional additional model-level properties, and other simulation tuning properties like schedulers or random number generators. The most common model type is StandardABM, but more specialized model types are also available, see AgentBasedModel.

(Trivial) evolve the model.

(Optional) Visualize the model and animate its time evolution. This can help checking that the model behaves as expected and there aren't any mistakes, or can be used in making figures for a paper/presentation.

Collect data. To do this, specify which data should be collected, by providing one standard Julia Vector of data-to-collect for agents, for example [:mood, :wealth], and another one for the model. The agent data names are given as the keyword adata and the model as keyword mdata to the function run!. This function outputs collected data in the form of a DataFrame.

If you're planning of running massive simulations, it might be worth having a look at the Performance Tips after familiarizing yourself with Agents.jl.

1. The space

Agents.jl offers several possibilities for the space the agents live in. In addition, it is straightforward to implement a fundamentally new type of space, see Creating a new space type.

The available spaces are listed in the Available spaces part of the API. An example of a space is OpenStreetMapSpace. It is based on Open Street Map, where agents are confined to move along streets of the map, using real-world values for the length of each street.

After deciding on the space, one simply initializes an instance of a space, e.g. with grid = GridSpace((10, 10)) and passes that into AgentBasedModel. See each individual space for all its possible arguments.

2. The agent type(s)

Agents in Agents.jl are instances of user-defined types. While the majority of Agents.jl API is based on a functional design, accessing agent properties is done with the simple field-access Julia syntax. For example, the (named) property weight of an agent can be obtained as agent.weight.

To create agent types, and define what properties they should have, the user needs to use the @agent macro, which ensures that agents have the minimum amount of required necessary properties to function within a given space and model by inheriting pre-defined agent properties suited for each type of space. The macro usage may seem intimidating at first, but it is in truth very simple! For example,

@agent struct Person(GridAgent{2})
    age::Int
    money::Float64
end

would make an agent type with named properties age, money, while also inheriting all named properties of the GridAgent{2} predefined type (which is necessary for simulating agents in a two-dimensional grid space).

Agents.@agent — Macro

@agent struct YourAgentType{X}(AnotherAgentType) [<: OptionalSupertype]
    extra_property::X
    other_extra_property_with_default::Bool = true
    const other_extra_const_property::Int
    # etc...
end

Define an agent struct which includes all fields that AnotherAgentType has, as well as any additional ones the user may provide. See below for examples.

Using @agent is the only supported way to create agent types for Agents.jl.

Structs created with @agent by default subtype AbstractAgent. They cannot subtype each other, as all structs created from @agent are concrete types and AnotherAgentType itself is also concrete (only concrete types have fields). If you want YourAgentType to subtype something other than AbstractAgent, use the optional argument OptionalSupertype (which itself must then subtype AbstractAgent).

Usage

The macro @agent has two primary uses:

To include the mandatory fields for a particular space in your agent struct. In this case you would use one of the minimal agent types as AnotherAgentType.

A convenient way to include fields from another, already existing struct.

The existing minimal agent types are:

NoSpaceAgent

GraphAgent

GridAgent

ContinuousAgent

OSMAgent

All will attribute an id::Int field, and besides NoSpaceAgent will also attribute a pos field. You should never directly manipulate the mandatory fields id, pos that the resulting new agent type will have. The id is an unchangeable field. Use functions like move_agent! etc., to change the position.

Examples

Example without optional hierarchy

Using

@agent struct Person{T}(GridAgent{2}) 
    age::Int
    moneyz::T
end

will create an agent appropriate for using with 2-dimensional GridSpace

mutable struct Person{T} <: AbstractAgent
    id::Int
    pos::NTuple{2, Int}
    const age::Int
    moneyz::T
end

Notice that you can also use default values for some fields, in this case you will need to specify the field names with the non-default values

@agent struct Person2{T}(GridAgent{2})
    age::Int = 30
    moneyz::T
end
# default age value
Person2(id = 1, pos = (1, 1), moneyz = 2000)
# new age value
Person2(1, (1, 1), 40, 2000)

Example with optional hierarchy

An alternative way to make the above structs, that also establishes a user-specific subtyping hierarchy would be to do:

abstract type AbstractHuman <: AbstractAgent end

@agent struct Worker(GridAgent{2}) <: AbstractHuman
    age::Int
    moneyz::Float64
end

@agent struct Fisher(Worker) <: AbstractHuman
    fish_per_day::Float64
end

which would now make both Fisher and Worker subtypes of AbstractHuman.

julia> supertypes(Fisher)
(Fisher, AbstractHuman, AbstractAgent, Any)

julia> supertypes(Worker)
(Worker, AbstractHuman, AbstractAgent, Any)

Note that Fisher will not be a subtype of Worker although Fisher has inherited the fields from Worker.

Example highlighting problems with parametric types

Notice that in Julia parametric types are union types. Hence, the following cannot be used:

@agent struct Dummy{T}(GridAgent{2})
    moneyz::T
end

@agent struct Fisherino{T}(Dummy{T})
    fish_per_day::T
end

You will get an error in the definition of Fisherino, because the fields of Dummy{T} cannot be obtained, because it is a union type. Same with using Dummy. You can only use Dummy{Float64}.

Example with common dispatch and no subtyping

It may be that you do not even need to create a subtyping relation if you want to utilize multiple dispatch. Consider the example:

@agent struct CommonTraits(GridAgent{2})
    age::Int
    speed::Int
    energy::Int
end

and then two more structs are made from these traits:

@agent struct Bird(CommonTraits)
    height::Float64
end

@agent struct Rabbit(CommonTraits)
    underground::Bool
end

If you wanted a function that dispatches to both Rabbit, Bird, you only have to define:

Animal = Union{Bird, Rabbit}
f(x::Animal) = ... # uses `CommonTraits` fields

However, it should also be said, that there is no real reason here to explicitly type-annotate x::Animal in f. Don't annotate any type. Annotating a type only becomes useful if there are at least two "abstract" groups, like Animal, Person. Then it would make sense to define

Person = Union{Fisher, Baker}
f(x::Animal) = ... # uses `CommonTraits` fields
f(x::Person) = ... # uses fields that all "persons" have

Agents.AbstractAgent — Type

YourAgentType <: AbstractAgent

Agents participating in Agents.jl simulations are instances of user-defined Types that are subtypes of AbstractAgent.

Your agent type(s) must have the id::Int field as first field. If any space is used (see Available spaces), a pos field of appropriate type is also mandatory. The core model structure, and each space, may also require additional fields that may, or may not, be communicated as part of the public API.

The @agent macro ensures that all of these constrains are in place and hence it is the the only supported way to create agent types.

3.1 The evolution rule - basics

The evolution rule may always be provided as one standard Julia function that inputs the model and modifies it in-place, according to the rules of the simulation. Such a model stepping function will itself likely call functions from the Agents.jl API and may look like

function model_step!(model)
    exchange = model.exchange # obtain the `exchange` model property
    agent = model[5] # obtain agent with ID = 5
    # Iterate over neighboring agents (within distance 1)
    for neighbor in nearby_agents(model, agent, 1)
        transfer = minimum(neighbor.money, exchange)
        agent.money += transfer
        neighbor.money -= transfer
    end
    return # function end. As it is in-place it `return`s nothing.
end

This function will be called once per simulation step.

As you can see, the above defined model stepping function did not operate on all agents of the model, only on agent with ID 5 and its spatial neighbors. Typically you would want to operate on more agents. There is a simple automated way to do this (section 3.2), and a non-automated, but fully-configurable way to do this (section 3.3).

3.2 The evolution rule - agent stepping function

In Agents.jl it is also possible to provide an agent stepping function. This feature enables scheduling agents automatically given some scheduling rule, skipping the agents that were scheduled to act but have been removed from the model (due to e.g., the actions of other agents), and also allows optimizations that are based on the specific type of AgentBasedModel.

An agent stepping function defines what happens to an agent when it activates. It inputs two arguments agent, model and operates in place on the agent and the model. This function will be applied to every agent that has been scheduled by the model's scheduler. A scheduler simply creates an iterator of agent IDs at each simulation step, possibly taking into account the current model state. Several schedulers are provided out-of-the-box by Agents.jl, see Schedulers.

Given the example above, an agent stepping function that would perform a similar currency exchange between agents would look like

function agent_step!(agent, model)
    exchange = model.exchange # obtain the `exchange` model property
    # Iterate over neighboring agents (within distance 1)
    for neighbor in nearby_agents(model, agent, 1)
        transfer = minimum(neighbor.money, exchange)
        agent.money += transfer
        neighbor.money -= transfer
    end
    agent.age += 1
    # if too old, pass fortune onto heir and remove from model
    if agent.age > 75
        heir = replicate!(agent, model)
        heir.age = 1
        remove_agent!(agent, modeL)
    end
    return # function end. As it is in-place it `return`s nothing.
end

and to activate all agents randomly once per simulation step, one would use Schedulers.Randomly() as the model scheduler.

We stress that in contrast to the above model_step!, this function will be called for every scheduled agent, while model_step! will only be called once per simulation step. Naturally, you may define both an agent and a model stepping functions. In this case the model stepping function would perform model-wide actions that are not limited to a particular agent.

3.3 The evolution rule - advanced (manual scheduling)

Some advanced models may require special handling for scheduling, or may need to schedule agents several times and act on different subsets of agents with different functions during a single simulation step. In such a scenario, it is more sensible to provide only a model stepping function, where all the dynamics is contained within.

Here is an example:

function complex_step!(model)
    # tip: these schedulers should be defined as properties of the model
    scheduler1 = Schedulers.Randomly()
    scheduler2 = user_defined_function_with_model_as_arg
    for id in schedule(model, scheduler1)
        agent_step1!(model[id], model)
    end
    intermediate_model_action!(model)
    for id in schedule(model, scheduler2)
        agent_step2!(model[id], model)
    end
    if model.step_counter % 100 == 0
        model_action_every_100_steps!(model)
    end
    final_model_action!(model)
    return
end

For defining your own schedulers, see Schedulers.

Current step number

Notice that the current step number is not explicitly given to the model stepping function, nor is contained in the model type, because this is useful only for a subset of ABMs. If you need the step information, implement this by adding a counting parameter into the model properties, and incrementing it by 1 each time the model stepping function is called.

4. The model

The ABM is an instance of a subtype of AgentBasedModel, most typically simply a StandardABM. A model is created by passing all inputs of steps 1-3 into the model constructor. Once the model is constructed, it can be populated by agents using the add_agent! function, which we highlight in the educative example below.

Agents.AgentBasedModel — Type

AgentBasedModel

An AgentBasedModel is the abstract supertype encompassing models in Agents.jl. All models are some concrete implementation of AgentBasedModel and follow its interface (see below). ABM is an alias to AgentBasedModel.

For backwards compatibility, the following function is valid:

AgentBasedModel(AgentType [, space]; properties, kwargs...) → model

which dispatches to StandardABM.

Available concrete implementations

StandardABM

UnkillableABM

Interface of AgentBasedModel

Here we the most important information on how to query an instance of AgentBasedModel:

model[id] gives the agent with given id.

abmproperties(model) gives the properties container stored in the model.

model.property: If the model properties is a dictionary with key type Symbol, or if it is a composite type (struct), then the syntax model.property will return the model property with key :property.

abmrng(model) will return the random number generator of the model. It is strongly recommended to use abmrng(model) to all calls to rand and similar functions, so that reproducibility can be established in your modelling workflow.

Many more functions exist in the API page, such as allagents.

Agents.StandardABM — Type

StandardABM <: AgentBasedModel

The most standard concrete implementation of an AgentBasedModel, as well as the default version of the generic AgentBasedModel constructor. StandardABM stores agents in a dictionary mapping unique Int IDs to agents. See also UnremovableABM for better performance in case number of agents can only increase during the model evolution.

StandardABM(AgentType [, space]; properties, kwargs...) → model

Creates a model expecting agents of type AgentType living in the given space. It can support supports multiple agent types by passing a Union of agent types as AgentType. Have a look at Performance Tips for potential drawbacks of this approach.

space is a subtype of AbstractSpace, see Space for all available spaces. If it is omitted then all agents are virtually in one position and there is no spatial structure. Spaces are mutable objects and are not designed to be shared between models. Create a fresh instance of a space with the same properties if you need to do this.

The evolution rules are functions given to the keywords agent_step!, model_step!, schedule. If agent_step! is not provided, the evolution rules is just the function given to model_step!. Each step of a simulation with StandardABM proceeds as follows: If agent_step! is not provided, then a simulation step is equivalent with calling model_step!. If agent_step! is provided, then a simulation step first schedules agents by calling the scheduler. Then, it applies the agent_step! function to all scheduled agents. Then, the model_step! function is called (optionally, the model_step! function may be called before activating the agents).

Keywords

agent_step! = dummystep: the optional stepping function for each agent contained in the model. For complicated models, it could be more suitable to use only model_step! to evolve the model.

model_step! = dummystep: the optional stepping function for the model.

properties = nothing: additional model-level properties that the user may decide upon and include in the model. properties can be an arbitrary container of data, however it is most typically a Dict with Symbol keys, or a composite type (struct).

scheduler = Schedulers.fastest: is the scheduler that decides the (default) activation order of the agents. See the scheduler API for more options.

rng = Random.default_rng(): the random number generation stored and used by the model in all calls to random functions. Accepts any subtype of AbstractRNG.

agents_first::Bool = true: whether to schedule and activate agents first and then call the model_step! function, or vice versa.

warn=true: some type tests for AgentType are done, and by default warnings are thrown when appropriate.

5. Evolving the model

After you have created an instance of an AgentBasedModel, it is rather trivial to evolve it by simply calling step! on it

CommonSolve.step! — Function

step!(model::ABM [, n::Int = 1])

Evolve the model for n steps according to the evolution rule.

step!(model, f::Function)

In this version, step! runs the model until f(model, s) returns true, where s is the current amount of steps taken, starting from 0.

See also Advanced stepping for stepping complex models where agent_step! might not be convenient.

5. Visualizations

Once you have defined a model and the stepping functions, you can visualize the model statically, or animate its time evolution straightforwardly in ~5 lines of code. This is discussed in a different page: Visualizations and Animations for Agent Based Models. Furthermore, all models in the Examples showcase plotting.

6. Collecting data

Running the model and collecting data while the model runs is done with the run! function. Besides run!, there is also the paramscan function that performs data collection while scanning ranges of the parameters of the model, and the ensemblerun! that performs ensemble simulations and data collection.

Agents.run! — Function

run!(model, n::Integer; kwargs...) → agent_df, model_df
run!(model, f::Function; kwargs...) → agent_df, model_df

Run the model (step it with the input arguments propagated into step!) and collect data specified by the keywords, explained one by one below. Return the data as two DataFrames, one for agent-level data and one for model-level data.

See also offline_run! to write data to file while running the model.

Data-deciding keywords

adata::Vector means "agent data to collect". If an entry is a Symbol, e.g. :weight, then the data for this entry is agent's field weight. If an entry is a Function, e.g. f, then the data for this entry is just f(a) for each agent a. The resulting dataframe columns are named with the input symbol (here :weight, :f).

adata::Vector{<:Tuple}: if adata is a vector of tuples instead, data aggregation is done over the agent properties.

For each 2-tuple, the first entry is the "key" (any entry like the ones mentioned above, e.g. :weight, f). The second entry is an aggregating function that aggregates the key, e.g. mean, maximum. So, continuing from the above example, we would have adata = [(:weight, mean), (f, maximum)].

It's also possible to provide a 3-tuple, with the third entry being a conditional function (returning a Bool), which assesses if each agent should be included in the aggregate. For example: x_pos(a) = a.pos[1]>5 with (:weight, mean, x_pos) will result in the average weight of agents conditional on their x-position being greater than 5.

The resulting data name columns use the function dataname. They create something like :mean_weight or :maximum_f_x_pos. In addition, you can use anonymous functions in a list comprehension to assign elements of an array into different columns: adata = [(a)->(a.interesting_array[i]) for i=1:N]. Column names can also be renamed with DataFrames.rename! after data is collected.

Notice: Aggregating only works if there are agents to be aggregated over. If you remove agents during model run, you should modify the aggregating functions. E.g. instead of passing mean, pass mymean(a) = isempty(a) ? 0.0 : mean(a).

mdata::Vector means "model data to collect" and works exactly like adata. For the model, no aggregation is possible (nothing to aggregate over).

Alternatively, mdata can also be a function. This is a "generator" function, that accepts model as input and provides a Vector that represents mdata. Useful in combination with an ensemblerun! call that requires a generator function.

By default both keywords are nothing, i.e. nothing is collected/aggregated.

Mixed-Models

For mixed-models, the adata keyword has some additional options & properties. An additional column agent_type will be placed in the output dataframe.

In the case that data is needed for one agent type that does not exist in a second agent type, missing values will be added to the dataframe.

Warning: Since this option is inherently type unstable, try to avoid this in a performance critical situation.

Aggregate functions will fail if missing values are not handled explicitly. If a1.weight but a2 (type: Agent2) has no weight, use a2(a) = a isa Agent2; adata = [(:weight, sum, a2)] to filter out the missing results.

Other keywords

when=true : at which steps s to perform the data collection and processing. A lot of flexibility is offered based on the type of when. If when::AbstractVector, then data are collected if s ∈ when. Otherwise data are collected if when(model, s) returns true. By default data are collected in every step.

when_model = when : same as when but for model data.

obtainer = identity : method to transfer collected data to the DataFrame. Typically only change this to copy if some data are mutable containers (e.g. Vector) which change during evolution, or deepcopy if some data are nested mutable containers. Both of these options have performance penalties.

showprogress=false : Whether to show progress

The run! function has been designed for maximum flexibility: nearly all scenarios of data collection are possible, whether you need agent data, model data, aggregated data, or arbitrary combinations.

Nevertheless, we also expose a simple data-collection API (see Data collection), that gives users even more flexibility, allowing them to make their own "data collection loops" arbitrarily calling step! and collecting data as, and when, needed.

As your models become more complex, it may not be advantageous to use lots of helper functions in the global scope to assist with data collection. If this is the case in your model, here's a helpful tip to keep things clean: use a generator function to collect data as instructed in the documentation string of run!. For example:

function assets(model)
    total_savings(model) = model.bank_balance + sum(model.assets)
    function strategy(model)
        if model.year == 0
            return model.initial_strategy
        else
            return get_strategy(model)
        end
    end
    return [:age, :details, total_savings, strategy]
end
run!(model, agent_step!, model_step!, 10; mdata = assets)

Seeding and Random numbers

Each ABM in Agents.jl contains a random number generator (RNG) instance that can be obtained with abmrng(model). A benefit of this approach is making models deterministic so that they can be run again and yield the same output. To do this, always pass a specifically seeded RNG to the model creation, e.g. rng = Random.MersenneTwister(1234) and then give this rng to the model creation.

For reproducibility and performance reasons, one should never use rand() without using the RNG, thus throughout our examples we use rand(abmrng(model)) or rand(abmrng(model), 1:10, 100), etc.

An educative example

A simple, education-oriented example of using the basic Agents.jl API is given in Schelling's segregation model.